Add handleValidations. Moved it to indep. folder
