+++ app.js
@@ -10,7 +10,7 @@ const formatsLogger = app.get("env") === "development" ? "dev" : "short";
 
 app.use(logger(formatsLogger));
 app.use(cors());
-app.use(express.json());
+app.use(express.json()); // Body parser middleware
 
 app.use("/api/contacts", contactsRouter);
 app.use((req, res) => {
