+++ models/contacts.json
@@ -22,11 +22,5 @@
     "name": "Antonio",
     "email": "vivaldi1444@gmail.com",
     "phone": "1234567890"
-  },
-  {
-    "id": "oebcVCjOqQMpGzMY3L4Ej",
-    "name": "Antonio Vivaldi",
-    "email": "vivaldi1444@gmail.com",
-    "phone": "1234567890"
   }
 ]
