+++ routes/api/contacts.js
@@ -6,44 +6,7 @@ const contacts = require("../../models/contacts");
 
 require("colors");
 
-async function isContactWithSameFields(res, email, phone) {
-  const contactsList = await contacts.listContacts();
-  const contact = Object.values(contactsList).some(
-    (contact) =>
-      [contact.email, contact.phone].includes(email) ||
-      [contact.email, contact.phone].includes(phone)
-  );
-
-  if (contact) {
-    res
-      .status(409)
-      .json({ message: "Contact with this email or phone already exists." })
-      .end();
-  }
-}
-async function isContactExist(res, id) {
-  const contact = await contacts.getContactById(id);
-
-  if (!contact || contact === undefined) {
-    res.status(404).json({ message: "Not found" }).end();
-  }
-  return contact;
-}
-function checkRequiredFields(res, name, email) {
-  const missedFields = [];
-
-  !name && missedFields.push("name");
-  !email && missedFields.push("email");
-  //*  !phone && missedFields.push("phone"); // maybe its not a required field
-
-  if (missedFields.length > 0) {
-    res
-      .status(400)
-      .json({ message: `Missing required fields: ${missedFields.join(", ")}` })
-      .end();
-  }
-}
-
+const handleValidations = require("../../controllers/handleValidations");
 router.get("/", async (req, res, next) => {
   // the route will be '3000:/api/contacts[get-path]
   try {
@@ -59,7 +22,7 @@ router.get("/:contactId", async (req, res, next) => {
   const { contactId } = req.params;
 
   try {
-    const data = await isContactExist(res, contactId);
+    const data = await handleValidations.isContactExist(res, contactId);
 
     res.send(data);
   } catch (error) {
@@ -72,8 +35,8 @@ router.post("/", async (req, res, next) => {
   const { name, email, phone } = req.body;
 
   try {
-    checkRequiredFields(res, name, email, phone);
-    await isContactWithSameFields(res, email, phone);
+    handleValidations.checkRequiredFields(res, name, email, phone);
+    await handleValidations.isContactWithSameFields(res, email, phone);
 
     const contact = await contacts.addContact(req.body);
     res.status(201).json(contact);
@@ -89,7 +52,7 @@ router.delete("/:contactId", async (req, res, next) => {
   try {
     const id = req.params.contactId;
 
-    await isContactExist(res, id);
+    await handleValidations.isContactExist(res, id);
     await contacts.removeContact(id);
 
     res.send("Contact is successfully removed!");
@@ -108,7 +71,7 @@ router.put("/:contactId", async (req, res, next) => {
     return;
   }
   try {
-    await isContactExist(res, id);
+    await handleValidations.isContactExist(res, id);
 
     const updatedContact = await contacts.updateContact(id, req.body);
